requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
idna==3.11
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyuca==1.2
requests==2.32.5
sentry-sdk==2.42.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
xlsxwriter==3.2.0
<<<<<<< HEAD
gunicorn==23.0.0

# testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
=======

gunicorn==23.0.0
>>>>>>> 996afe303a3a4caade0c07ac7f12d4c403c82b67
